Start refactoring the index monolith
This introduces a script loader, which automatically runs and executes files in the /scripts/ folder.

This allows us to refactor all of the scripts into individual files.

As an example, we refactor the on.ready function which previously logged the user tag and then loaded all of the commands into two separate scripts